Add haddock and clean up warnings

+++ src/Data/Stuffed.hs
@@ -8,6 +8,26 @@
 {-# LANGUAGE DataKinds                  #-}
 {-# LANGUAGE KindSignatures             #-}
 
+-- | This module tries to be as efficient as possible, type safe and easy to use. If you have a "sink" like
+--
+-- > sink :: Stuffed 0 -> IO ()
+-- > sink = undefined
+--
+-- You'd then simply be able to encode a Bytestring with
+--
+-- > sink $ stuff bytes
+--
+-- You can try this out in ghci with:
+--
+-- >>> :set -XOverloadedStrings
+-- >>> :set -XDataKinds
+-- >>> import Data.Stuffed
+-- >>> let stuffedBytes = stuff "a\0b\0c" :: Stuffed 0
+-- >>> unpack $ unwrap stuffedBytes -- directly access the underlying bytestring
+-- [2,97,2,98,2,99]
+-- >>> unpack $ unstuff stuffedBytes
+-- [97,0,98,0,99]
+
 module Data.Stuffed
     ( Stuffed
     , stuff
@@ -27,10 +47,11 @@ import           Data.Semigroup           (Semigroup)
 import           Data.Word                (Word8)
 import           GHC.Generics             (Generic)
 import           GHC.Types                (Nat)
-import           Prelude                  hiding (concat, length, null, splitAt)
+import           Prelude                  hiding (concat, length, null, splitAt, last)
 
 import           Data.Stuffed.Internal    (IsByte)
 
+-- | Wrapper for Lazy Bytestrings, parametrized on the Byte (Word8, represented as a type-level 'Nat') to be encoded away.
 newtype Stuffed (a :: Nat) = Stuffed ByteString
     deriving (Eq, Ord, Show, Semigroup, Monoid, Generic)
 
@@ -49,7 +70,7 @@ stuff bs = Stuffed $ fromChunks chunks
 buildStuffed :: Word8 -> B.ByteString -> B.ByteString
 buildStuffed excluded bs = B.cons last stuffed
     where
-        excludedOffset = fromIntegral excluded + 1
+        excludedOffset = excluded + 1
         swapExcluded n current | current == excluded = (excludedOffset, n)
                                | otherwise = (n + 1, current)
         (last, stuffed) = B.mapAccumR swapExcluded excludedOffset bs
@@ -68,5 +89,6 @@ rebuildChunk excluded (Just (starting_offset, rest)) _ = byteString $ snd $ B.ma
         swapExcluded n current | n == excluded = (current - 1, excluded)
                                | otherwise = (n - 1, current)
 
+-- | Extract the encoded bytestring from a Stuffed
 unwrap :: Stuffed a -> ByteString
 unwrap (Stuffed bs) = bs
